requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
black==23.7.0
blip-ci==0.0.5
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
clip-interrogator==0.5.4
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.3
fairscale==0.4.4
fastapi==0.101.1
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gradio==3.40.1
gradio_client==0.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.2
open-clip-torch==2.20.0
orjson==3.9.4
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
protobuf==3.20.3
pydantic==2.1.1
pydantic_core==2.4.0
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.9.2
safetensors==0.3.2
semantic-version==2.10.0
sentencepiece==0.1.99
simple-image-download==0.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
timm==0.4.12
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.26.1
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.4
uvicorn==0.23.2
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2
